+++ lib/src/platform/file_picker_platform_interface.dart
@@ -3,6 +3,8 @@ import 'dart:typed_data';
 
 import 'package:file_picker/src/api/file_picker_result.dart';
 import 'package:file_picker/src/api/file_picker_types.dart';
+import 'package:file_picker/src/api/android_saf_options.dart';
+import 'package:file_picker/src/api/file_picker_options.dart';
 import 'package:file_picker/src/platform/file_picker_method_channel.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
@@ -39,8 +41,7 @@ abstract class FilePickerPlatform extends PlatformInterface {
     bool withData = false,
     bool withReadStream = false,
     bool lockParentWindow = false,
-    bool readSequential = false,
-    bool cancelUploadOnWindowBlur = true,
+    FilePickerOptions options = const FilePickerOptions(),
   }) async {
     throw UnimplementedError('pickFiles() has not been implemented.');
   }
@@ -57,6 +58,11 @@ abstract class FilePickerPlatform extends PlatformInterface {
     );
   }
 
+  /// Releases the given SAF grant.
+  Future<void> releaseSAFGrant(String uri) async {
+    throw UnimplementedError('releaseSAFGrant() has not been implemented.');
+  }
+
   /// Asks the underlying platform to remove any temporary files created by this plugin.
   Future<bool?> clearTemporaryFiles() async {
     throw UnimplementedError('clearTemporaryFiles() has not been implemented.');
@@ -67,6 +73,7 @@ abstract class FilePickerPlatform extends PlatformInterface {
     String? dialogTitle,
     bool lockParentWindow = false,
     String? initialDirectory,
+    AndroidSAFOptions? androidSafOptions,
   }) async {
     throw UnimplementedError('getDirectoryPath() has not been implemented.');
   }