+++ lib/src/platform/web/file_picker_web.dart
@@ -5,6 +5,8 @@ import 'dart:typed_data';
 import 'package:file_picker/src/api/file_picker_types.dart';
 import 'package:file_picker/src/api/file_picker_result.dart';
 import 'package:file_picker/src/api/platform_file.dart';
+import 'package:file_picker/src/api/android_saf_options.dart';
+import 'package:file_picker/src/api/file_picker_options.dart';
 import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:path/path.dart' as p;
@@ -39,19 +41,28 @@ class FilePickerWeb extends FilePickerPlatform {
   }
 
   @override
+  Future<String?> getDirectoryPath({
+    String? dialogTitle,
+    bool lockParentWindow = false,
+    String? initialDirectory,
+    AndroidSAFOptions? androidSafOptions,
+  }) async {
+    throw UnimplementedError('getDirectoryPath() has not been implemented.');
+  }
+
+  @override
   Future<FilePickerResult?> pickFiles({
     String? dialogTitle,
     String? initialDirectory,
     FileType type = FileType.any,
     List<String>? allowedExtensions,
-    bool allowMultiple = false,
     Function(FilePickerStatus)? onFileLoading,
+    int compressionQuality = 0,
+    bool allowMultiple = false,
     bool withData = true,
     bool withReadStream = false,
     bool lockParentWindow = false,
-    bool readSequential = false,
-    bool cancelUploadOnWindowBlur = true,
-    int compressionQuality = 0,
+    FilePickerOptions options = const FilePickerOptions(),
   }) async {
     if (type != FileType.custom && (allowedExtensions?.isNotEmpty ?? false)) {
       throw Exception(
@@ -147,7 +158,7 @@ class FilePickerWeb extends FilePickerPlatform {
           syncCompleter.complete();
         });
         reader.readAsArrayBuffer(file);
-        if (readSequential) {
+        if (options.webOptions.readSequential) {
           await syncCompleter.future;
         }
       }
@@ -171,7 +182,7 @@ class FilePickerWeb extends FilePickerPlatform {
     uploadInput.addEventListener('change', changeEventListener.toJS);
     uploadInput.addEventListener('cancel', cancelledEventListener.toJS);
 
-    if (cancelUploadOnWindowBlur) {
+    if (options.webOptions.cancelUploadOnWindowBlur) {
       // Listen focus event for cancelled
       window.addEventListener('focus', cancelledEventListener.toJS);
     }